+++ RELEASE-NOTES.md
@@ -1,10 +1,11 @@
 # govueadmin.g8 Release Notes
 
-## 2020-09-11: template-v0.2.0
+## 2020-09-xx: template-v0.2.0
 
 - Use [Gravatar](https://gravatar.com/) for user profile picture.
 - AB#21: Migrate to CoreUI-vue v3.1.1
-- Vuejs: replace `data` with `computed` (https://techformist.com/data-vs-computed-vs-watcher-in-vue-components/).
+- AB#25: Use JWT as login-token
+- Vuejs: replace `data` with `computed` (https://techformist.com/data-vs-computed-vs-watcher-in-vue-components/) if possible.
 - Other fixes and enhancements.